Organize random_test.cpp and table.cpp into more files
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/rahulmalik87/pstress?shareId=XXXX-XXXX-XXXX-XXXX).
