+++ src/CMakeLists.txt
@@ -7,7 +7,7 @@ IF (MYSQL_FOUND)
   ELSE()
     INCLUDE_DIRECTORIES( ${MYSQL_INCLUDE_DIR} )
   ENDIF(MARIADB)
-  ADD_EXECUTABLE(${BINARY_NAME}-${PSTRESS_EXT} pstress.cpp help.cpp node.cpp thread.cpp random_test.cpp json.cpp table.cpp)
+  ADD_EXECUTABLE(${BINARY_NAME}-${PSTRESS_EXT} pstress.cpp help.cpp node.cpp thread.cpp random_test.cpp json.cpp table.cpp table_creation.cpp column_generation.cpp index_management.cpp query_execution.cpp table_operations.cpp bulk_inserts.cpp column_values.cpp table_serialization.cpp)
   TARGET_LINK_LIBRARIES( ${BINARY_NAME}-${PSTRESS_EXT} ${MYSQL_LIBRARY} ${OTHER_LIBS} inih++)
 
  # Copy all .sql files
