+++ src/query_execution.cpp
@@ -1,0 +1,36 @@
+#include "query_execution.hpp"
+#include "random_test.hpp"
+#include <iostream>
+#include <string>
+#include <vector>
+
+bool Thd1::run_some_query() {
+  // Implementation of the run_some_query function
+  // This function will execute some queries and return true if successful
+  // Add your query execution logic here
+  return true;
+}
+
+bool Thd1::load_metadata() {
+  // Implementation of the load_metadata function
+  // This function will load metadata and return true if successful
+  // Add your metadata loading logic here
+  return true;
+}
+
+bool execute_sql(const std::string &sql, Thd1 *thd, bool log_result_client) {
+  // Implementation of the execute_sql function
+  // This function will execute the given SQL query and return true if successful
+  // Add your SQL execution logic here
+  return true;
+}
+
+void print_and_log(std::string &&str, Thd1 *thd, bool print_error) {
+  // Implementation of the print_and_log function
+  // This function will print and log the given string
+  // Add your logging logic here
+  std::cout << str << std::endl;
+  if (print_error) {
+    std::cerr << str << std::endl;
+  }
+}
